# Code refactoring for ranking objective & Faster ndcg_xendcg