--- a/src/objective/objective_function.cpp
+++ b/src/objective/objective_function.cpp
@@ -7,7 +7,6 @@
 #include "binary_objective.hpp"
 #include "multiclass_objective.hpp"
 #include "rank_objective.hpp"
-#include "rank_xendcg_objective.hpp"
 #include "regression_objective.hpp"
 #include "xentropy_objective.hpp"