--- a/tests/python_package_test/test_sklearn.py
+++ b/tests/python_package_test/test_sklearn.py
@@ -129,7 +129,7 @@ class TestSklearn(unittest.TestCase):
                 eval_metric='ndcg',
                 callbacks=[lgb.reset_parameter(learning_rate=lambda x: max(0.01, 0.1 - 0.01 * x))])
         self.assertLessEqual(gbm.best_iteration_, 24)
-        self.assertGreater(gbm.best_score_['valid_0']['ndcg@1'], 0.6559)
+        self.assertGreater(gbm.best_score_['valid_0']['ndcg@1'], 0.6382)
         self.assertGreater(gbm.best_score_['valid_0']['ndcg@3'], 0.6421)
 
     def test_regression_with_custom_objective(self):