1. optimize the O(n^2) loop in `ndcg_xendcg` to O(n).
2. Fix non-thread-safe random generator in `ndcg_xendcg`, which should not share a `Random` object in multi-threading. Furthermore, the random results should be identical now regards to different num_threads.

Also some code refactorings.

Benchmark:


Boosting cost:
| Data      | Master|  This PR |
|----|  ----| ----| 
| Yahoo LTR | 9.36 s |1.45 s |
| MS LTR|72.02 s | 2.72 s|