@guolinke 2020-02-23
ping @sbruch for the changes in `ndcg_xendcg` part