@StrikerRUS 2020-03-15
@guolinke This PR removes old params related to ranking objective. Do you think it is not "breaking"?